Initial pronterface config
Config file for Pronterface where macros are stored. Within is "macro LASER_TEST_PATN", in which a grid of lines 50mm in length spaced by 1mm is created. The file itself, printrunconf.ini, must live in the C:\Users\your_user\ folder. New macros created will exist in this file and can be called in pronterface by typing the macro name in the pronterface terminal.